# Fix Rust UB problems