--- a/rust/flatbuffers/src/verifier.rs
+++ b/rust/flatbuffers/src/verifier.rs
@@ -232,22 +232,6 @@ impl<'opts, 'buf> Verifier<'opts, 'buf> {
         self.num_tables = 0;
         self.num_tables = 0;
     }
-    /// Check that there really is a T in there.
-    #[inline]
-    fn is_aligned<T>(&self, pos: usize) -> Result<()> {
-        Ok(())
-        // // Safe because we're not dereferencing.
-        // let p = unsafe { self.buffer.as_ptr().add(pos) };
-        // if (p as usize) % std::mem::align_of::<T>() == 0 {
-        //     Ok(())
-        // } else {
-        //     Err(InvalidFlatbuffer::Unaligned {
-        //         unaligned_type: std::any::type_name::<T>(),
-        //         position: pos,
-        //         error_trace: Default::default(),
-        //     })
-        // }
-    }
     #[inline]
     fn range_in_buffer(&mut self, pos: usize, size: usize) -> Result<()> {
         let end = pos.saturating_add(size);
@@ -262,7 +246,6 @@ impl<'opts, 'buf> Verifier<'opts, 'buf> {
     }
     #[inline]
     pub fn in_buffer<T>(&mut self, pos: usize) -> Result<()> {
-        self.is_aligned::<T>(pos)?;
         self.range_in_buffer(pos, std::mem::size_of::<T>())
     }
     #[inline]
@@ -314,7 +297,6 @@ impl<'opts, 'buf> Verifier<'opts, 'buf> {
     ) -> Result<TableVerifier<'ver, 'opts, 'buf>> {
         let vtable_pos = self.deref_soffset(table_pos)?;
         let vtable_len = self.get_u16(vtable_pos)? as usize;
-        self.is_aligned::<VOffsetT>(vtable_pos.saturating_add(vtable_len))?; // i.e. vtable_len is even.
         self.range_in_buffer(vtable_pos, vtable_len)?;
         // Check bounds.
         self.num_tables += 1;
@@ -463,7 +445,6 @@ impl<T: Verifiable> Verifiable for ForwardsUOffset<T> {
 fn verify_vector_range<T>(v: &mut Verifier, pos: usize) -> Result<std::ops::Range<usize>> {
     let len = v.get_uoffset(pos)? as usize;
     let start = pos.saturating_add(SIZE_UOFFSET);
-    v.is_aligned::<T>(start)?;
     let size = len.saturating_mul(std::mem::size_of::<T>());
     let end = start.saturating_add(size);
     v.range_in_buffer(start, size)?;