--- a/tests/RustTest.sh
+++ b/tests/RustTest.sh
@@ -50,3 +50,6 @@ else
 fi
 
 cargo bench $TARGET_FLAG
+
+rustup +nightly component add miri
+cargo miri test -- -Zmiri-disable-isolation