--- a/tests/RustTest.sh
+++ b/tests/RustTest.sh
@@ -52,7 +52,7 @@ fi
 cargo bench $TARGET_FLAG
 
 # RUST_NIGHTLY environment variable set in dockerfile.
-if [[ -z RUST_NIGHTLY ]]; then
+if [[ $RUST_NIGHTLY == 1 ]]; then
   rustup +nightly component add miri
   cargo +nightly miri test -- -Zmiri-disable-isolation
 fi