Hi @rw, @aardappel, @krojew  

This should fix #6359, #6203, #5854, #5825, and the last issue mentioned in #4916.

## Motivation

Miri is a rust interpreter that detects bad memory activity, i.e. UB. I've fixed the detected problems, which turned out to be alignment issues, and added it to `RustTest.sh`. With the verifier and UB issues sorted (so far as a core Rust tool can detect) I think the library is now acceptably safe according to the common standards of the Rust community.

## Details

I didn't run all tests through Miri. Its incredibly slow so I've disabled it for fuzzing and other expensive tests.
`verifier_one_byte_errors_do_not_crash` is disabled but I did run it manually since its quite important. That took like an hour on my weak little macbook air. Miri tests are only running on linux since it tests an abstract memory model and there shouldn't be any difference on other platforms.

Since Rust's `Vec<u8>` and `[u8]` only promises alignment 1, we have to copy out of the buffer and onto the stack to get aligned access. This is a very minor performance hit that's incurred when scalars are accessed. The alternative would be to introduce a nonstandard type such as `Vec<u128>` or even a custom vector that promises 16byte alignment. That's pretty inconvenient for us and users.

I had to rework the implementation of structs in rust to use an array of bytes rather than `repr(C, align=4)`. This is because we return references to structs from fb-vectors, and these references are alignment 1. Please take a closer look at the new implementation.

Alignment checks from the verifier are now with respect to `buffer[0]` as per the Flatbuffers specification. They may not be aligned w.r.t. the system because of `[u8]`'s alignment. I kept this check in case a verified buffer in Rust is assumed aligned when passed to another language.