@aardappel 2021-01-08
I am pretty surprised Rust doesn't have a built-in way to have aligned buffers of `u8`. That is pretty fundamental for any serialized data, file formats, memory mapping etc. No Rust allocator will align at 1 byte granularity so it isn't even a practical issue.

There seems to be a lot in Rust that favors being pedantic over pragmatism, that is sad.