@CasperN 2021-01-08
> No Rust allocator will align at 1 byte granularity so it isn't even a practical issue.

Rust by default uses jemalloc and it does actually take advantage of lower alignment requirements to reduce fragmentation. This is why we have users observing alignment errors (#6359).

> I am pretty surprised Rust doesn't have a built-in way to have aligned buffers of u8

I mean you can get aligned raw pointers via `std::alloc` but its lower level than what most people prefer to use.

There'll probably be a way to parameterize `Vec<u8>` with an allocator that guarantees minimum 16byte alignment, and eventually maybe even a way for us to restrict allowable allocators via the type system. That doesn't exist yet but there's a working group talking about it.

https://github.com/rust-lang/wg-allocators/issues

The biggest issue is `&[u8]`. That's a very fundamental and widely used type and I don't see a good way for `&[u8]` promise a minimum alignment since you can take arbitrary subslices.