@rw 2021-01-10
@CasperN Nice work!

What are the benchmarks like with these changes? I wonder if representing structs as arrays will cause the compiler to use the stack more than it should (instead of registers). And I'm curious about the effects of the copying that @krojew mentioned.