@CasperN 2021-01-10
@rw

> What are the benchmarks like with these changes? 

There's some performance regression on my machine, but its not too significant, within the "+/- interval."

Master:
```
running 16 tests
test create_byte_vector_100_naive       ... bench:         856 ns/iter (+/- 93) = 116 MB/s
test create_byte_vector_100_optimal     ... bench:          28 ns/iter (+/- 4) = 3571 MB/s
test create_canonical_buffer_then_reset ... bench:       1,096 ns/iter (+/- 170) = 277 MB/s
test create_string_10                   ... bench:          35 ns/iter (+/- 25) = 285 MB/s
test create_string_100                  ... bench:          35 ns/iter (+/- 7) = 2828 MB/s
test traverse_canonical_buffer          ... bench:         128 ns/iter (+/- 35) = 2375 MB/s
```

This branch:
```
running 16 tests
test create_byte_vector_100_naive       ... bench:         907 ns/iter (+/- 181) = 110 MB/s
test create_byte_vector_100_optimal     ... bench:          29 ns/iter (+/- 4) = 3448 MB/s
test create_canonical_buffer_then_reset ... bench:       1,120 ns/iter (+/- 173) = 267 MB/s
test create_string_10                   ... bench:          30 ns/iter (+/- 6) = 333 MB/s
test create_string_100                  ... bench:          36 ns/iter (+/- 6) = 2750 MB/s
test traverse_canonical_buffer          ... bench:         127 ns/iter (+/- 25) = 2362 MB/s
```

> I wonder if representing structs as arrays will cause the compiler to use the stack more than it should (instead of registers). 

Maybe, but I wouldn't look into it. The regression is very tolerable, given we're eliminating UB.

Anything to address before merging?