@CasperN 2021-01-12
> WDYT about us running miri for all tests sometimes?

It might be good idea but I don't know how long it takes, Miri seems around 5 orders of magnitude slower than normal rust.