@CasperN 2021-01-15
>  WDYT about us marking generated functions as unsafe, if they use unsafe? To @eduardosm 's point, it seems like users could potentially think that it's okay to call those functions themselves, when in general we don't want them to.

Yea that seems like a simple solution to this